src/pd/message.rs: Translate from original ZenGarden C++ source
Adds the `heapless` crate as a backing store for `pd::Message` elements.

Several parse errors which were previously ignored are now treated as
error cases which return a `Result`. This may or may not reflect the
requirements, but this code is starting out by being pessimistic about
error handling, with the goal of only relaxing it if necessary.