src/message/object.rs: Translate from C++ (and add Allocators)
This commit translates the `message::Object` code into Rust.

One big problem with the existing C++ code which maps poorly to Rust is
its use of copious mutable references everywhere for representing
connections between object. Additionally, it makes heavy use of
subclasses for modelling objects.

To avoid these problems, this commit also begins work to migrate the
structure of the codebase from an OO system into something more like an
Entity-Component-System as described in the blog post
"My RustConf 2018 Closing Keynote":

https://kyren.github.io/2018/09/14/rustconf-talk.html

This pattern not only works exceptionally well for this particular
library, whose behavior is not too far off from a game engine,
but also provides a great answer to the question of how we allocate
and store objects:

This commit also introduces a `FixedSizedAllocator` type which contains
a typed allocator with a fixed (via typenum / const generics) number
of slots, and a system for tracking objects with IDs and obtaining
references to them by ID.