src/pd/context.rs: Translate from C++ (also message sends, queue)
This commit translates the `pd::Context` code into Rust.

This is the core domain object and, under an ECS model, the primary
holder of all state. Since we don't have all of the types this uses yet,
large parts of it are commented out, but have hopefully been translated
to semi-sane rust.

It also translates `message::Send`, `message::send::Controller`, and
`message::OrderedQueue` as these types are important to the core message
delivery behavior, so they've been translated in advance.

A few placeholder types were used which need to get replaced eventually:

- `pd::Graph`, `pd::graph::Id`
- `message::send::controller::Set`,
  `message::send::controller::RemoteMessageReceiver`